Update ai_commands.js
This should fix #96

+++ src/ai_commands.js
@@ -281,6 +281,9 @@ const changeConsoleParams = () => {
 };
 
 const killScript = () => {
+  if (!runner)
+    return;
+
   runner.stdin.pause();
   runner.kill();
 };
